move bitcoin websocket to its own directory
