+++ packages/blockchain-wallet-v4/src/redux/middleware.js
@@ -1,5 +1,5 @@
 import walletSync from './walletSync/middleware.js'
-import socket from './webSocket/middleware.js'
+import * as socket from './webSocket/middleware.js'
 import kvStore from './kvStore/middleware.js'
 
 export {
