+++ packages/blockchain-wallet-v4/src/redux/webSocket/actionTypes.js
@@ -1,6 +1,5 @@
-export const START_SOCKET = '@CORE.START_SOCKET'
-export const STOP_SOCKET = '@CORE.STOP_SOCKET'
+import * as bitcoin from './bitcoin/actionTypes'
 
-export const OPEN_SOCKET = '@CORE.OPEN_SOCKET'
-export const MESSAGE_SOCKET = '@CORE.MESSAGE_SOCKET'
-export const CLOSE_SOCKET = '@CORE.CLOSE_SOCKET'
+export {
+  bitcoin
+}
