+++ packages/blockchain-wallet-v4/src/redux/webSocket/actions.js
@@ -1,8 +1,5 @@
-import * as T from './actionTypes'
+import * as bitcoin from './bitcoin/actions'
 
-export const startSocket = () => ({ type: T.START_SOCKET })
-export const stopSocket = () => ({ type: T.STOP_SOCKET })
-
-export const openSocket = () => ({ type: T.OPEN_SOCKET })
-export const messageSocket = (payload) => ({ type: T.MESSAGE_SOCKET, payload })
-export const closeSocket = () => ({ type: T.CLOSE_SOCKET })
+export {
+  bitcoin
+}
