+++ packages/blockchain-wallet-v4/src/redux/webSocket/bitcoin/actionTypes.js
@@ -1,0 +1,6 @@
+export const START_SOCKET = '@CORE.BITCOIN.START_SOCKET'
+export const STOP_SOCKET = '@CORE.BITCOIN.STOP_SOCKET'
+
+export const OPEN_SOCKET = '@CORE.BITCOIN.OPEN_SOCKET'
+export const MESSAGE_SOCKET = '@CORE.BITCOIN.MESSAGE_SOCKET'
+export const CLOSE_SOCKET = '@CORE.BITCOIN.CLOSE_SOCKET'
